Completely removed bottom navigation and all fragments
