+++ app/src/main/java/com/example/bookreservations/MainActivity.java
@@ -3,40 +3,60 @@ package com.example.bookreservations;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.os.Build;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.MotionEvent;
+import android.view.View;
 import android.view.inputmethod.InputMethodManager;
+import android.widget.Button;
+import android.widget.ImageView;
 
-import com.google.android.material.bottomnavigation.BottomNavigationView;
 
+import androidx.annotation.RequiresApi;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.navigation.NavController;
-import androidx.navigation.Navigation;
-import androidx.navigation.ui.AppBarConfiguration;
-import androidx.navigation.ui.NavigationUI;
 import androidx.preference.PreferenceManager;
 
-public class MainActivity extends AppCompatActivity implements SharedPreferences.OnSharedPreferenceChangeListener {
+import com.android.volley.RequestQueue;
+import com.android.volley.toolbox.Volley;
+import com.example.bookreservations.utils.JsonAPIparser;
+import com.example.bookreservations.utils.Timer;
 
+public class MainActivity extends AppCompatActivity implements SharedPreferences.OnSharedPreferenceChangeListener {
+    String apiUrl;
+    JsonAPIparser parser;
+    View root;
+
+    @RequiresApi(api = Build.VERSION_CODES.M)
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
         setContentView(R.layout.activity_main);
-        BottomNavigationView navView = findViewById(R.id.nav_view);
+        root = findViewById(android.R.id.content).getRootView();
 
-        // Passing each menu ID as a set of Ids because each
-        // menu should be considered as top level destinations.
-        AppBarConfiguration appBarConfiguration = new AppBarConfiguration.Builder(
-                R.id.navigation_home, R.id.navigation_about, R.id.navigation_notifications)
-                .build();
-        NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment);
-        NavigationUI.setupActionBarWithNavController(this, navController, appBarConfiguration);
-        NavigationUI.setupWithNavController(navView, navController);
+        ImageView my_image = findViewById(R.id.FirstImage);
+        my_image.setImageResource(R.drawable.logo_knihovna);
 
+        Button refresh_button = findViewById(R.id.refresh_button);
+        refresh_button.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                parser.jsonParse();
+            }
+        });
+
+        setupSharedPreferences();
+
+        RequestQueue mQueue = Volley.newRequestQueue(this);
+        parser = new JsonAPIparser(this, root, mQueue, apiUrl);
+        new Timer(root, parser);
+
+        if (savedInstanceState == null) {
+            parser.jsonParse();
+        }
 
         setupSharedPreferences();
     }
@@ -71,13 +91,26 @@ public class MainActivity extends AppCompatActivity implements SharedPreferences
 
     private void setupSharedPreferences() {
         SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
+        this.apiUrl = sharedPreferences.getString("api_url", null);
+
         sharedPreferences.registerOnSharedPreferenceChangeListener(this);
     }
 
 
+    @RequiresApi(api = Build.VERSION_CODES.M)
     @Override
     public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
+        if (key.equals("api_url")) {
 
+            this.apiUrl = sharedPreferences.getString("api_url", null);
+
+            //setTextVisible(sharedPreferences.getBoolean("display_text",true));
+            // Should also set view and buttons that timer uses
+            RequestQueue mQueue = Volley.newRequestQueue(this);
+            parser = new JsonAPIparser(this, root, mQueue, apiUrl);
+            new Timer(root, parser);
+
+        }
     }
 
     @Override
