--- app/src/main/java/com/example/bookreservations/ui/about/AboutViewModel.java
@@ -1,19 +1,0 @@
-package com.example.bookreservations.ui.about;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class AboutViewModel extends ViewModel {
-
-    private MutableLiveData<String> mText;
-
-    public AboutViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is about fragment");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
