--- app/src/main/java/com/example/bookreservations/ui/home/HomeViewModel.java
@@ -1,21 +1,0 @@
-package com.example.bookreservations.ui.home;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class HomeViewModel extends ViewModel {
-
-    private MutableLiveData<String> mText;
-
-    public HomeViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
-
-
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
