--- app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:id="@+id/navigation_home"
-        android:icon="@drawable/ic_home_black_24dp"
-        android:title="@string/title_home" />
-
-    <item
-        android:id="@+id/navigation_notifications"
-        android:icon="@drawable/ic_notifications_black_24dp"
-        android:title="@string/title_notifications" />
-    <item
-        android:id="@+id/navigation_about"
-        android:icon="@drawable/ic_dashboard_black_24dp"
-        android:title="@string/title_about" />
-
-</menu>
