+++ app/src/main/res/menu/settings_menu.xml
@@ -5,7 +5,7 @@
     <item
         android:id="@+id/action_settings"
         android:title="@string/action_settings"
-        android:orderInCategory="100"
+        android:orderInCategory="101"
         app:showAsAction="never"
         />
