+++ app/src/main/res/values/colors.xml
@@ -1,6 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <color name="colorPrimary">#FF5722</color>
-    <color name="colorPrimaryDark">#D64011</color>
-    <color name="colorAccent">#FFC107</color>
-</resources>
+<resources></resources>
